get de produtos para fazer pedido feito

+++ src/main/java/webb_lanches/webb_lanches/Produtos/DTO/ProdutoDTO.java
@@ -1,7 +1,7 @@
 package webb_lanches.webb_lanches.Produtos.DTO;
 
 public record ProdutoDTO(
-    int idProduto,
+    Long idProduto,
     String nomeProduto,
     double preco,
     String descricao,
